chore: accept base
